Refactor where PAL_meta is stored
A proposal to refactor where PAL_meta information is stored. The idea
is to have enough meta data to be able to recreate a configuration
file that produced the DataTree and to have a system that is flexible
enough to handle PalProcesses that may mutate data inplace or take
multiple DataSets as input.

Each PalProcess now contains a 'output_dataset' parameter that can
be set in the process' configuration and stores the path to the DataSet
it operates on. PAL_meta attributes are set on the output dataset that
is a dictionary with keys equal to process names and configurations
as values.

The root Group of a DataTree will have a PAL_meta attribute containing
a map with 'ouput_datasets' that is a list of all DataSets in the
DataTree that contain PAL_meta attributes.